% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acquire_license}
\alias{acquire_license}
\title{Acquire an IAI license for the current session.}
\usage{
acquire_license(...)
}
\arguments{
\item{...}{Refer to the Julia documentation for available parameters}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/installation/#IAI.acquire_license}{\code{IAI.acquire_license}}
}
\section{IAI Compatibility}{

Requires IAI version 3.1 or higher.
}

\examples{
\dontrun{iai::acquire_license()}

}
