% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_julia_processes}
\alias{add_julia_processes}
\title{Add additional Julia worker processes to parallelize workloads}
\usage{
add_julia_processes(...)
}
\arguments{
\item{...}{Refer to the Julia documentation for available parameters}
}
\description{
Julia Equivalent:
\href{https://docs.julialang.org/en/v1/stdlib/Distributed/#Distributed.addprocs}{\code{Distributed.addprocs!}}
}
\details{
For more information, refer to the
\href{https://docs.interpretable.ai/stable/IAIBase/advanced/#IAIBase-Parallelization-1}{documentation on parallelization}
}
\examples{
\dontrun{iai::add_julia_processes(3)}

}
