% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{all_treatment_combinations}
\alias{all_treatment_combinations}
\title{Return a dataframe containing all treatment combinations of one or more
treatment vectors, ready for use as treatment candidates in `fit_predict!`
or `predict`}
\usage{
all_treatment_combinations(...)
}
\arguments{
\item{...}{A vector of possible options for each treatment}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/RewardEstimation/reference/#IAI.all_treatment_combinations}{\code{IAI.all_treatment_combinations}}
}
\examples{
\dontrun{iai::all_treatment_combinations(c(1, 2, 3))}

}
