% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{apply}
\alias{apply}
\title{Return the leaf index in a tree model into which each point in the features
falls}
\usage{
apply(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAITrees/reference/#IAI.apply}{\code{IAI.apply}}
}
\examples{
\dontrun{iai::apply(lnr, X)}

}
