% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{apply_nodes}
\alias{apply_nodes}
\title{apply_nodes}
\usage{
apply_nodes(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{X}{The features of the data.}
}
\description{
Return the indices of the points in the features that fall into each node of
a trained tree model
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.apply_nodes}{\code{IAI.apply_nodes}}
}
\examples{
\dontrun{iai::apply_nodes(lnr, X)}

}
