% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{autoplot.grid_search}
\alias{autoplot.grid_search}
\title{Construct a
\href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}}
object plotting grid search results for Optimal Feature Selection learners}
\usage{
\method{autoplot}{grid_search}(object, type = stop("`type` is required"), ...)
}
\arguments{
\item{object}{The grid search to plot}

\item{type}{The type of plot to construct (either \code{"validation"} or
\code{"importance"}, for more information refer to the
\href{https://docs.interpretable.ai/v3.2.1/OptimalFeatureSelection/visualization/#Plotting-Grid-Search-Results-1}{Julia documentation for plotting grid search results}
)}

\item{...}{Additional arguments (unused)}
}
\description{
Construct a
\href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}}
object plotting grid search results for Optimal Feature Selection learners
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{ggplot2::autoplot(grid)}

}
