% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{autoplot.roc_curve}
\alias{autoplot.roc_curve}
\title{Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the ROC curve}
\usage{
\method{autoplot}{roc_curve}(x, ...)
}
\arguments{
\item{x}{The ROC curve to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the ROC curve
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{ggplot2::autoplot(roc)}

}
