% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{autoplot.similarity_comparison}
\alias{autoplot.similarity_comparison}
\title{Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the results of the similarity comparison}
\usage{
\method{autoplot}{similarity_comparison}(x, ...)
}
\arguments{
\item{x}{The similarity comparison to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the results of the similarity comparison
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{ggplot2::autoplot(similarity)}

}
