% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{autoplot.stability_analysis}
\alias{autoplot.stability_analysis}
\title{Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the results of the stability analysis}
\usage{
\method{autoplot}{stability_analysis}(object, ...)
}
\arguments{
\item{object}{The stability analysis to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Construct a \href{https://ggplot2.tidyverse.org/reference/ggplot.html}{\code{ggplot2::ggplot}} object plotting the results of the stability analysis
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{ggplot2::autoplot(stability)}

}
