% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{categorical_regression_reward_estimator}
\alias{categorical_regression_reward_estimator}
\title{Learner for conducting reward estimation with categorical treatments and
regression outcomes}
\usage{
categorical_regression_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/RewardEstimation/reference/#IAI.CategoricalRegressionRewardEstimator}{\code{IAI.CategoricalRegressionRewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{lnr <- iai::categorical_regression_reward_estimator()}

}
