% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{categorical_reward_estimator}
\alias{categorical_reward_estimator}
\title{Learner for conducting reward estimation with categorical treatments}
\usage{
categorical_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
This function was deprecated in iai 1.6.0, and
[categorical_classification_reward_estimator()] or
[categorical_classification_reward_estimator()] should be used instead.
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{lnr <- iai::categorical_reward_estimator()}

}
