% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{categorical_survival_reward_estimator}
\alias{categorical_survival_reward_estimator}
\title{Learner for conducting reward estimation with categorical treatments and
survival outcomes}
\usage{
categorical_survival_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/RewardEstimation/reference/#IAI.CategoricalSurvivalRewardEstimator}{\code{IAI.CategoricalSurvivalRewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{lnr <- iai::categorical_survival_reward_estimator()}

}
