% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{cleanup_installation}
\alias{cleanup_installation}
\title{Remove all traces of automatic Julia/IAI installation}
\usage{
cleanup_installation()
}
\description{
Removes files created by \code{\link{install_julia}} and
\code{\link{install_system_image}}
}
\examples{
\dontrun{iai::cleanup_installation()}

}
