% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{convert_treatments_to_numeric}
\alias{convert_treatments_to_numeric}
\title{Convert `treatments` from symbol/string format into numeric values.}
\usage{
convert_treatments_to_numeric(treatments)
}
\arguments{
\item{treatments}{The treatments to convert}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/RewardEstimation/reference/#IAI.convert_treatments_to_numeric}{\code{IAI.convert_treatments_to_numeric}}
}
\examples{
\dontrun{iai::convert_treatments_to_numeric(c("1", "2", "3"))}

}
