% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{decision_path}
\alias{decision_path}
\title{decision_path}
\usage{
decision_path(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{X}{The features of the data.}
}
\description{
Return a matrix where entry \code{(i, j)} is true if the \code{i}th point in
the features passes through the \code{j}th node in a trained tree model.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.decision_path}{\code{IAI.decision_path}}
}
\examples{
\dontrun{iai::decision_path(lnr, X)}

}
