% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{delete_rich_output_param}
\alias{delete_rich_output_param}
\title{Delete a global rich output parameter}
\usage{
delete_rich_output_param(key)
}
\arguments{
\item{key}{The parameter to delete.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.delete_rich_output_param!}{\code{IAI.delete_rich_output_param!}}
}
\examples{
\dontrun{iai::delete_rich_output_param("simple_layout")}

}
