% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{equal_propensity_estimator}
\alias{equal_propensity_estimator}
\title{Learner that estimates equal propensity for all treatments.}
\usage{
equal_propensity_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
For use with data from randomized experiments where treatments are known to
be randomly assigned.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/RewardEstimation/reference/#IAI.EqualPropensityEstimator}{\code{IAI.EqualPropensityEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::equal_propensity_estimator()}

}
