% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{fit.imputation_learner}
\alias{fit.imputation_learner}
\title{Fits an imputation learner to the training data.}
\usage{
\method{fit}{imputation_learner}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to fit.}

\item{X}{The features of the data.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Additional keyword arguments are available for fitting imputation learners -
please refer to the Julia documentation.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/OptImpute/reference/#IAI.fit\%21-Tuple\%7BImputationLearner\%7D}{\code{IAI.fit!}}
}
\examples{
\dontrun{iai::fit(lnr, X)}

}
