% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{fit.optimal_feature_selection_learner}
\alias{fit.optimal_feature_selection_learner}
\title{Fits an Optimal Feature Selection learner to the training data}
\usage{
\method{fit}{optimal_feature_selection_learner}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to fit.}

\item{X}{The features of the data.}

\item{...}{Other parameters, including zero or more target vectors as
required by the problem type. Refer to the Julia documentation for
available parameters.}
}
\description{
When the \code{coordinated_sparsity} parameter of the learner is \code{TRUE},
additional keyword arguments are required - please refer to the Julia
documentation.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/OptimalFeatureSelection/reference/#IAI.fit\%21-Tuple\%7BOptimalFeatureSelectionLearner\%7D}{\code{IAI.fit!}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::fit(lnr, X)}

}
