% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{fit_and_expand}
\alias{fit_and_expand}
\title{Fit an imputation learner with training features and create adaptive
indicator features to encode the missing pattern}
\usage{
fit_and_expand(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner to use for imputation.}

\item{X}{The dataframe in which to impute missing values.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/OptImpute/reference/#IAI.fit_and_expand!}{\code{IAI.fit_and_expand!}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{lnr <- iai::fit_and_expand(lnr, X, type = "finite")}

}
