% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_best_params}
\alias{get_best_params}
\title{get_best_params}
\usage{
get_best_params(grid)
}
\arguments{
\item{grid}{The grid search to query.}
}
\description{
Return the best parameter combination from a grid
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.get_best_params}{\code{IAI.get_best_params}}
}
\examples{
\dontrun{iai::get_best_params(grid)}

}
