% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_classification_proba}
\alias{get_classification_proba}
\title{Generic function for returning the probabilities of class membership at a
node of a classification tree}
\usage{
get_classification_proba(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the probabilities of class membership at a
node of a classification tree
}
