% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_depth}
\alias{get_depth}
\title{Get the depth of a node of a tree}
\usage{
get_depth(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAITrees/reference/#IAI.get_depth}{\code{IAI.get_depth}}
}
\examples{
\dontrun{iai::get_depth(lnr, 1)}

}
