% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_features_used}
\alias{get_features_used}
\title{Return the names of the features used by the learner}
\usage{
get_features_used(lnr)
}
\arguments{
\item{lnr}{The learner to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.get_features_used}{\code{IAI.get_features_used}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::get_features_used(lnr)}

}
