% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_grid_result_details}
\alias{get_grid_result_details}
\title{Return a vector of lists detailing the results of the grid search}
\usage{
get_grid_result_details(grid)
}
\arguments{
\item{grid}{The grid search to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/IAIBase/reference/#IAI.get_grid_result_details}{\code{IAI.get_grid_result_details}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::get_grid_result_details(grid)}

}
