% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_grid_result_summary}
\alias{get_grid_result_summary}
\title{Return a summary of the results from the grid search}
\usage{
get_grid_result_summary(grid)
}
\arguments{
\item{grid}{The grid search to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAIBase/reference/#IAI.get_grid_result_summary}{\code{IAI.get_grid_result_summary}}
}
\examples{
\dontrun{iai::get_grid_result_summary(grid)}

}
