% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_grid_results}
\alias{get_grid_results}
\title{Return a summary of the results from the grid search}
\usage{
get_grid_results(grid)
}
\arguments{
\item{grid}{The grid search to query.}
}
\description{
This function was deprecated and renamed to [get_grid_result_summary()]
in iai 2.1.0. This is for consistency with the IAI v2.2.0 Julia release.
}
\examples{
\dontrun{iai::get_grid_results(grid)}

}
