% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_lower_child}
\alias{get_lower_child}
\title{Get the index of the lower child at a split node of a tree}
\usage{
get_lower_child(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_lower_child}{\code{IAI.get_lower_child}}
}
\examples{
\dontrun{iai::get_lower_child(lnr, 1)}

}
