% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_machine_id}
\alias{get_machine_id}
\title{Return the machine ID for the current computer.}
\usage{
get_machine_id()
}
\description{
This ID ties the IAI license file to your machine.
}
\examples{
\dontrun{iai::get_machine_id()}

}
