% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{get_num_fits.glmnetcv_learner}
\alias{get_num_fits.glmnetcv_learner}
\title{Return the number of fits along the path in a trained GLMNet learner}
\usage{
\method{get_num_fits}{glmnetcv_learner}(obj, ...)
}
\arguments{
\item{obj}{The GLMNet learner to query.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/Heuristics/reference/#IAI.get_num_fits-Tuple\%7BGLMNetCVLearner\%7D}{\code{IAI.get_num_fits}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::get_num_fits(lnr)}

}
