% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{get_num_fits.optimal_feature_selection_learner}
\alias{get_num_fits.optimal_feature_selection_learner}
\title{Return the number of fits along the path in a trained Optimal Feature
Selection learner}
\usage{
\method{get_num_fits}{optimal_feature_selection_learner}(obj, ...)
}
\arguments{
\item{obj}{The Optimal Feature Selection learner to query.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/OptimalFeatureSelection/reference/#IAI.get_num_fits-Tuple\%7BOptimalFeatureSelectionLearner\%7D}{\code{IAI.get_num_fits}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{iai::get_num_fits(lnr)}

}
