% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_num_fits}
\alias{get_num_fits}
\title{Generic function for returning the number of fits in a trained learner}
\usage{
get_num_fits(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the number of fits in a trained learner
}
