% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_num_nodes}
\alias{get_num_nodes}
\title{Return the number of nodes in a trained learner}
\usage{
get_num_nodes(lnr)
}
\arguments{
\item{lnr}{The learner to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.get_num_nodes}{\code{IAI.get_num_nodes}}
}
\examples{
\dontrun{iai::get_num_nodes(lnr)}

}
