% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_params}
\alias{get_params}
\title{Return the value of all parameters on a learner}
\usage{
get_params(lnr)
}
\arguments{
\item{lnr}{The learner to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.get_params}{\code{IAI.get_params}}
}
\examples{
\dontrun{iai::get_params(lnr)}

}
