% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_parent}
\alias{get_parent}
\title{Get the index of the parent node at a node of a tree}
\usage{
get_parent(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAITrees/reference/#IAI.get_parent}{\code{IAI.get_parent}}
}
\examples{
\dontrun{iai::get_parent(lnr, 2)}

}
