% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_policy_treatment_outcome}
\alias{get_policy_treatment_outcome}
\title{Return the quality of the treatments at a node of a tree}
\usage{
get_policy_treatment_outcome(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.get_policy_treatment_outcome}{\code{IAI.get_policy_treatment_outcome}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::get_policy_treatment_outcome(lnr, 1)}

}
