% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_policy_treatment_outcome_standard_error}
\alias{get_policy_treatment_outcome_standard_error}
\title{Return the standard error for the quality of the treatments at a node of a
tree}
\usage{
get_policy_treatment_outcome_standard_error(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.get_policy_treatment_outcome_standard_error}{\code{IAI.get_policy_treatment_outcome_standard_error}}
}
\section{IAI Compatibility}{

Requires IAI version 3.2 or higher.
}

\examples{
\dontrun{iai::get_policy_treatment_outcome_standard_error(lnr, 1)}

}
