% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{get_prediction_constant.glmnetcv_learner}
\alias{get_prediction_constant.glmnetcv_learner}
\title{Return the constant term in the prediction in a trained
GLMNet learner}
\usage{
\method{get_prediction_constant}{glmnetcv_learner}(obj, fit_index = NULL, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{fit_index}{The index of the fit in the path to use for prediction,
defaulting to the best fit if not supplied.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/Heuristics/reference/#IAI.get_prediction_constant-Tuple\%7BGLMNetCVLearner\%7D}{\code{IAI.get_prediction_constant}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::get_prediction_constant(lnr)}

}
