% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_prediction_constant}
\alias{get_prediction_constant}
\title{Generic function for returning the prediction constant in a trained learner}
\usage{
get_prediction_constant(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the prediction constant in a trained learner
}
