% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{get_prediction_weights.glmnetcv_learner}
\alias{get_prediction_weights.glmnetcv_learner}
\title{Return the weights for numeric and categoric features used for prediction in
a trained GLMNet learner}
\usage{
\method{get_prediction_weights}{glmnetcv_learner}(obj, fit_index = NULL, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{fit_index}{The index of the fit in the path to use for prediction,
defaulting to the best fit if not supplied.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/Heuristics/reference/#IAI.get_prediction_weights-Tuple\%7BGLMNetCVLearner\%7D}{\code{IAI.get_prediction_weights}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::get_prediction_weights(lnr)}

}
