% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{get_prediction_weights.optimal_feature_selection_learner}
\alias{get_prediction_weights.optimal_feature_selection_learner}
\title{Return the weights for numeric and categoric features used for prediction in
a trained Optimal Feature Selection learner}
\usage{
\method{get_prediction_weights}{optimal_feature_selection_learner}(obj, fit_index = NULL, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{fit_index}{The index of the cluster to use for prediction, if the
\code{coordinated_sparsity} parameter on the learner is
\code{TRUE}.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/OptimalFeatureSelection/reference/#IAI.get_prediction_weights-Tuple\%7BOptimalFeatureSelectionLearner\%7D}{\code{IAI.get_prediction_weights}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::get_prediction_weights(lnr)}

}
