% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_regression_constant}
\alias{get_regression_constant}
\title{Generic function for returning the constant term in the regression
prediction at a node of a tree}
\usage{
get_regression_constant(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the constant term in the regression
prediction at a node of a tree
}
