% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_regression_weights.prescription_tree_learner}
\alias{get_regression_weights.prescription_tree_learner}
\title{Return the weights for each feature in the linear regression prediction at a
node of a prescription tree}
\usage{
\method{get_regression_weights}{prescription_tree_learner}(obj, node_index, treatment, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{treatment}{The treatment to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.get_regression_weights-Tuple\%7BPrescriptionTreeLearner\%2C\%20Int64\%2C\%20Any\%7D}{\code{IAI.get_regression_weights}}
}
\examples{
\dontrun{iai::get_regression_weights(lnr, 1, "A")}

}
