% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_regression_weights}
\alias{get_regression_weights}
\title{Return the weights for each feature in the regression prediction at a node of
a tree}
\usage{
get_regression_weights(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{If a prescription problem, the treatment to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/}{\code{IAI.get_regression_weights}}
(for regression or prescription tree learners as appropriate)
}
\examples{
\dontrun{
iai::get_regression_weights(lnr, 1)
iai::get_regression_weights(lnr, 1, "A")
}

}
