% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_regression_weights.regression_tree_learner}
\alias{get_regression_weights.regression_tree_learner}
\title{Return the weights for each feature in the linear regression prediction at a
node of a regression tree}
\usage{
\method{get_regression_weights}{regression_tree_learner}(obj, node_index, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.get_regression_weights-Tuple\%7BRegressionTreeLearner\%2C\%20Int64\%7D}{\code{IAI.get_regression_weights}}
}
\examples{
\dontrun{iai::get_regression_weights(lnr, 1)}

}
