% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_rich_output_params}
\alias{get_rich_output_params}
\title{Return the current global rich output parameter settings}
\usage{
get_rich_output_params()
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.get_rich_output_params}{\code{IAI.get_rich_output_params}}
}
\examples{
\dontrun{iai::get_rich_output_params()}

}
