% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_split_categories}
\alias{get_split_categories}
\title{Return the categoric/ordinal information used in the split at a node of a
tree}
\usage{
get_split_categories(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAITrees/reference/#IAI.get_split_categories}{\code{IAI.get_split_categories}}
}
\examples{
\dontrun{iai::get_split_categories(lnr, 1)}

}
