% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_split_feature}
\alias{get_split_feature}
\title{Return the feature used in the split at a node of a tree}
\usage{
get_split_feature(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_split_feature}{\code{IAI.get_split_feature}}
}
\examples{
\dontrun{iai::get_split_feature(lnr, 1)}

}
