% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_split_weights}
\alias{get_split_weights}
\title{get_split_weights}
\usage{
get_split_weights(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Return the weights for numeric and categoric features used in the hyperplane
split at a node of a tree
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_split_weights}{\code{IAI.get_split_weights}}
}
\examples{
\dontrun{iai::get_split_weights(lnr, 1)}

}
