% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_stability_results}
\alias{get_stability_results}
\title{Return the trained trees in order of increasing objective value,
along with their variable importance scores for each feature}
\usage{
get_stability_results(stability)
}
\arguments{
\item{stability}{The stability analysis to query}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAITrees/reference/#IAI.get_stability_results}{\code{get_stability_results}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::get_stability_results(stability)}

}
