% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_survival_expected_time}
\alias{get_survival_expected_time}
\title{Return the predicted expected survival time at a node of a tree}
\usage{
get_survival_expected_time(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.get_survival_expected_time}{\code{IAI.get_survival_expected_time}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::get_survival_expected_time(lnr, 1)}

}
