% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_survival_hazard}
\alias{get_survival_hazard}
\title{Return the predicted hazard ratio at a node of a tree}
\usage{
get_survival_hazard(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAITrees/reference/#IAI.get_survival_hazard}{\code{IAI.get_survival_hazard}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::get_survival_hazard(lnr, 1)}

}
