% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_train_errors}
\alias{get_train_errors}
\title{Extract the training objective value for each candidate tree in the
comparison, where a lower value indicates a better solution}
\usage{
get_train_errors(similarity)
}
\arguments{
\item{similarity}{The similarity comparison}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAITrees/reference/#IAI.get_train_errors}{\code{get_train_errors}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::get_train_errors(similarity)}

}
