% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_tree}
\alias{get_tree}
\title{Return a copy of the learner that uses a specific tree rather than the tree
with the best training objective.}
\usage{
get_tree(lnr, index)
}
\arguments{
\item{lnr}{The original learner}

\item{index}{The index of the tree to use}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAITrees/reference/#IAI.get_tree}{\code{get_tree}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::get_tree(lnr, index)}

}
