% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{glmnetcv_classifier}
\alias{glmnetcv_classifier}
\title{Learner for training GLMNet models for classification problems with
cross-validation}
\usage{
glmnetcv_classifier(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/Heuristics/reference/#IAI.GLMNetCVClassifier}{\code{IAI.GLMNetCVClassifier}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{lnr <- iai::glmnetcv_classifier()}

}
