% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{glmnetcv_regressor}
\alias{glmnetcv_regressor}
\title{Learner for training GLMNet models for regression problems with
cross-validation}
\usage{
glmnetcv_regressor(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/Heuristics/reference/#IAI.GLMNetCVRegressor}{\code{IAI.GLMNetCVRegressor}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::glmnetcv_regressor()}

}
