% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{glmnetcv_survival_learner}
\alias{glmnetcv_survival_learner}
\title{Learner for training GLMNet models for survival problems with
cross-validation}
\usage{
glmnetcv_survival_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/Heuristics/reference/#IAI.GLMNetCVSurvivalLearner}{\code{IAI.GLMNetCVSurvivalLearner}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{lnr <- iai::glmnetcv_survival_learner()}

}
