% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{grid_search}
\alias{grid_search}
\title{grid_search}
\usage{
grid_search(lnr, ...)
}
\arguments{
\item{lnr}{The learner to use when validating.}

\item{...}{The parameters to validate over.}
}
\description{
Controls grid search over parameter combinations
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.GridSearch}{\code{IAI.GridSearch}}
}
\examples{
\donttest{
iai::iai_setup()
grid <- iai::grid_search(
    iai::optimal_tree_classifier(
        random_seed = 1,
    ),
    max_depth = 1:5,
)
}
}
