% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{imputation_learner}
\alias{imputation_learner}
\title{Generic learner for imputing missing values}
\usage{
imputation_learner(method = "opt_knn", ...)
}
\arguments{
\item{method}{(optional) Specifies the imputation method to use.}

\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/OptImpute/reference/#IAI.ImputationLearner}{\code{IAI.ImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::imputation_learner(method = "opt_tree")}

}
