% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(X, ...)
}
\arguments{
\item{X}{The dataframe in which to impute missing values.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Impute missing values using either a specified method or through validation
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.impute}{\code{IAI.impute}}
}
\examples{
\donttest{
X <- iris
X[1, 1] <- NA
iai::iai_setup()
iai::impute(X)
}
}
