% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{is_categoric_split}
\alias{is_categoric_split}
\title{Check if a node of a tree applies a categoric split}
\usage{
is_categoric_split(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.is_categoric_split}{\code{IAI.is_categoric_split}}
}
\examples{
\dontrun{iai::is_categoric_split(lnr, 1)}

}
