% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{is_hyperplane_split}
\alias{is_hyperplane_split}
\title{Check if a node of a tree applies a hyperplane split}
\usage{
is_hyperplane_split(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAITrees/reference/#IAI.is_hyperplane_split}{\code{IAI.is_hyperplane_split}}
}
\examples{
\dontrun{iai::is_hyperplane_split(lnr, 1)}

}
