% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{is_leaf}
\alias{is_leaf}
\title{is_leaf}
\usage{
is_leaf(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Check if a node of a tree is a leaf
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.is_leaf}{\code{IAI.is_leaf}}
}
\examples{
\dontrun{iai::is_leaf(lnr, 1)}

}
