% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_graphviz}
\alias{load_graphviz}
\title{Loads the Julia Graphviz library to permit certain visualizations.}
\usage{
load_graphviz()
}
\description{
The library will be installed if not already present.
}
\examples{
\dontrun{iai::load_graphviz()}

}
