% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{mean_imputation_learner}
\alias{mean_imputation_learner}
\title{Learner for conducting mean imputation}
\usage{
mean_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/OptImpute/reference/#IAI.MeanImputationLearner}{\code{IAI.MeanImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::mean_imputation_learner()}

}
