% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{missing_goes_lower}
\alias{missing_goes_lower}
\title{Check if points with missing values go to the lower child at a split node of
of a tree}
\usage{
missing_goes_lower(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.missing_goes_lower}{\code{IAI.missing_goes_lower}}
}
\examples{
\dontrun{iai::missing_goes_lower(lnr, 1)}

}
