% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{multi_questionnaire.default}
\alias{multi_questionnaire.default}
\title{Construct an interactive questionnaire using multiple tree learners as
specified by questions}
\usage{
\method{multi_questionnaire}{default}(obj, ...)
}
\arguments{
\item{obj}{The questions to visualize. Refer to the \href{https://docs.interpretable.ai/v2.0.0/IAITrees/visualization/#multivis-1}{Julia documentation on multi-learner visualizations} for more information.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAITrees/reference/#IAI.MultiQuestionnaire}{\code{IAI.MultiQuestionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{
iai::multi_questionnaire(list("Questionnaire for" = list(
   "first learner" = lnr1,
   "second learner" = lnr2
)))
}

}
