% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{multi_questionnaire}
\alias{multi_questionnaire}
\title{Generic function for constructing an interactive questionnaire with
multiple learners}
\usage{
multi_questionnaire(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for constructing an interactive questionnaire with
multiple learners
}
