% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{multi_tree_plot.grid_search}
\alias{multi_tree_plot.grid_search}
\title{Construct an interactive tree visualization of multiple tree learners from
the results of a grid search}
\usage{
\method{multi_tree_plot}{grid_search}(obj, ...)
}
\arguments{
\item{obj}{The grid to visualize}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAITrees/reference/#IAI.MultiTreePlot-Tuple\%7BGridSearch\%7D}{\code{IAI.MultiTreePlot}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::multi_tree_plot(grid)}

}
