% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{multi_tree_plot}
\alias{multi_tree_plot}
\title{Generic function for constructing an interactive tree visualization of
multiple tree learners}
\usage{
multi_tree_plot(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for constructing an interactive tree visualization of
multiple tree learners
}
