% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{numeric_reward_estimator}
\alias{numeric_reward_estimator}
\title{Learner for conducting reward estimation with numeric treatments}
\usage{
numeric_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
This function was deprecated in iai 1.6.0, and
[numeric_classification_reward_estimator()] or
[numeric_classification_reward_estimator()] should be used instead.
}
\details{
This deprecation is no longer supported as of the IAI v3 release.
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or 2.2.
}

\examples{
\dontrun{lnr <- iai::numeric_reward_estimator()}

}
