% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{numeric_survival_reward_estimator}
\alias{numeric_survival_reward_estimator}
\title{Learner for conducting reward estimation with numeric treatments and survival
outcomes}
\usage{
numeric_survival_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/RewardEstimation/reference/#IAI.NumericSurvivalRewardEstimator}{\code{IAI.NumericSurvivalRewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{lnr <- iai::numeric_survival_reward_estimator()}

}
