% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{opt_knn_imputation_learner}
\alias{opt_knn_imputation_learner}
\title{Learner for conducting optimal k-NN imputation}
\usage{
opt_knn_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.OptKNNImputationLearner}{\code{IAI.OptKNNImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::opt_knn_imputation_learner()}

}
