% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{opt_svm_imputation_learner}
\alias{opt_svm_imputation_learner}
\title{Learner for conducting optimal SVM imputation}
\usage{
opt_svm_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/OptImpute/reference/#IAI.OptSVMImputationLearner}{\code{IAI.OptSVMImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::opt_svm_imputation_learner()}

}
