% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{opt_tree_imputation_learner}
\alias{opt_tree_imputation_learner}
\title{opt_tree_imputation_learner}
\usage{
opt_tree_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Learner for conducting optimal tree-based imputation
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.OptTreeImputationLearner}{\code{IAI.OptTreeImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::opt_tree_imputation_learner()}

}
