% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_classifier}
\alias{optimal_tree_classifier}
\title{Learner for training Optimal Classification Trees}
\usage{
optimal_tree_classifier(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/OptimalTrees/reference/#IAI.OptimalTreeClassifier}{\code{IAI.OptimalTreeClassifier}}
}
\examples{
\dontrun{lnr <- iai::optimal_tree_classifier()}

}
