% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_prescription_maximizer}
\alias{optimal_tree_prescription_maximizer}
\title{Learner for training Optimal Prescriptive Trees where the prescriptions
should aim to maximize outcomes}
\usage{
optimal_tree_prescription_maximizer(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/OptimalTrees/reference/#IAI.OptimalTreePrescriptionMaximizer}{\code{IAI.OptimalTreePrescriptionMaximizer}}
}
\examples{
\dontrun{lnr <- iai::optimal_tree_prescription_maximizer()}

}
