% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_prescription_minimizer}
\alias{optimal_tree_prescription_minimizer}
\title{Learner for training Optimal Prescriptive Trees where the prescriptions
should aim to minimize outcomes}
\usage{
optimal_tree_prescription_minimizer(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/OptimalTrees/reference/#IAI.OptimalTreePrescriptionMinimizer}{\code{IAI.OptimalTreePrescriptionMinimizer}}
}
\examples{
\dontrun{lnr <- iai::optimal_tree_prescription_minimizer()}

}
