% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_survivor}
\alias{optimal_tree_survivor}
\title{optimal_tree_survivor}
\usage{
optimal_tree_survivor(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Learner for training Optimal Survival Trees
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptimalTrees/stable/reference/#IAI.OptimalTreeSurvivor}{\code{IAI.OptimalTreeSurvivor}}
}
\examples{
\dontrun{lnr <- iai::optimal_tree_survivor()}

}
