% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{plot.grid_search}
\alias{plot.grid_search}
\title{Plot a grid search results for Optimal Feature Selection learners}
\usage{
\method{plot}{grid_search}(x, ...)
}
\arguments{
\item{x}{The grid search to plot}

\item{...}{Additional arguments (passed to
\href{https://docs.interpretable.ai/v2.2.0/IAI-R/reference/#ggplot2::autoplot.grid_search}{\code{ggplot2::autoplot.grid_search}})}
}
\description{
Plot a grid search results for Optimal Feature Selection learners
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{plot(grid)}

}
