% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{plot.roc_curve}
\alias{plot.roc_curve}
\title{Plot an ROC curve}
\usage{
\method{plot}{roc_curve}(x, ...)
}
\arguments{
\item{x}{The ROC curve to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Plot an ROC curve
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{plot(roc)}

}
