% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{plot.similarity_comparison}
\alias{plot.similarity_comparison}
\title{Plot a similarity comparison}
\usage{
\method{plot}{similarity_comparison}(x, ...)
}
\arguments{
\item{x}{The similarity comparison to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Plot a similarity comparison
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{plot(similarity)}

}
