% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{plot.stability_analysis}
\alias{plot.stability_analysis}
\title{Plot a stability analysis}
\usage{
\method{plot}{stability_analysis}(x, ...)
}
\arguments{
\item{x}{The stability analysis to plot}

\item{...}{Additional arguments (unused)}
}
\description{
Plot a stability analysis
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{plot(stability)}

}
