% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{predict.numeric_reward_estimator}
\alias{predict.numeric_reward_estimator}
\title{Return counterfactual rewards estimated by a numeric reward estimator for
each observation in the supplied data}
\usage{
\method{predict}{numeric_reward_estimator}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to use for estimation}

\item{X}{The features of the data.}

\item{...}{Additional arguments depending on the treatment and outcome types.
Refer to the Julia documentation for more information.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/RewardEstimation/reference/#IAI.predict-Tuple\%7BNumericRewardEstimator\%7D}{\code{IAI.predict}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::predict(lnr, X, treatments, outcomes)}

}
