% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Return the predictions made by the model for each point in the features
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.predict}{\code{IAI.predict}}
}
\examples{
\dontrun{iai::predict(lnr, X)}

}
