% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict.supervised_learner}
\alias{predict.supervised_learner}
\title{Return the predictions made by a supervised learner for each point in the
features}
\usage{
\method{predict}{supervised_learner}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/IAIBase/reference/#IAI.predict-Tuple\%7BSupervisedLearner\%2C\%20Union\%7BDataFrames.AbstractDataFrame\%2C\%20AbstractMatrix\%7B\%3C\%3AReal\%7D\%7D\%7D}{\code{IAI.predict}}
}
\examples{
\dontrun{iai::predict(lnr, X)}

}
