% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict_expected_survival_time}
\alias{predict_expected_survival_time}
\title{Generic function for returning the expected survival time predicted by a
model}
\usage{
predict_expected_survival_time(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the expected survival time predicted by a
model
}
