% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_expected_survival_time.survival_curve}
\alias{predict_expected_survival_time.survival_curve}
\title{Return the expected survival time estimate made by a survival curve
(as returned by
\code{\link{predict.survival_learner}}
or
\code{\link{get_survival_curve}})}
\usage{
\method{predict_expected_survival_time}{survival_curve}(obj, ...)
}
\arguments{
\item{obj}{The survival curve to use for prediction.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.predict_expected_survival_time-Tuple\%7BSurvivalCurve\%7D}{\code{IAI.predict_expected_survival_time}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::predict_expected_survival_time(curve)}

}
