% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict_hazard}
\alias{predict_hazard}
\title{Generic function for returning the hazard coefficient predicted by a model}
\usage{
predict_hazard(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the hazard coefficient predicted by a model
}
