% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_outcomes.policy_learner}
\alias{predict_outcomes.policy_learner}
\title{Return the predicted outcome for each treatment made by a policy
learner for each point in the features}
\usage{
\method{predict_outcomes}{policy_learner}(obj, X, rewards, ...)
}
\arguments{
\item{obj}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{rewards}{The estimated reward matrix for the data.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAIBase/reference/#IAI.predict_outcomes-Tuple\%7BPolicyLearner\%2C\%20Union\%7BDataFrames.AbstractDataFrame\%2C\%20AbstractMatrix\%7B\%3C\%3AReal\%7D\%7D\%2C\%20Union\%7BDataFrames.AbstractDataFrame\%2C\%20AbstractMatrix\%7B\%3C\%3AReal\%7D\%7D\%7D}{\code{IAI.predict_outcomes}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher
}

\examples{
\dontrun{iai::predict_outcomes(lnr, X, rewards)}

}
