% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_outcomes.prescription_learner}
\alias{predict_outcomes.prescription_learner}
\title{Return the predicted outcome for each treatment made by a prescription
learner for each point in the features}
\usage{
\method{predict_outcomes}{prescription_learner}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.predict_outcomes-Tuple\%7BPrescriptionLearner\%2C\%20Union\%7BDataFrames.AbstractDataFrame\%2C\%20AbstractMatrix\%7B\%3C\%3AReal\%7D\%7D\%7D}{\code{IAI.predict_outcomes}}
}
\examples{
\dontrun{iai::predict_outcomes(lnr, X)}

}
