% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_outcomes}
\alias{predict_outcomes}
\title{Return the the predicted outcome for each treatment made by a model for each
point in the features}
\usage{
predict_outcomes(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.predict_outcomes}{\code{IAI.predict_outcomes}}
}
\examples{
\dontrun{iai::predict_outcomes(lnr, X)}

}
