% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_proba}
\alias{predict_proba}
\title{predict_proba}
\usage{
predict_proba(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Return the probabilities of class membership predicted by a model for each
point in the features
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.predict_proba}{\code{IAI.predict_proba}}
}
\examples{
\dontrun{iai::predict_proba(lnr, X)}

}
