% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{predict_reward.categorical_reward_estimator}
\alias{predict_reward.categorical_reward_estimator}
\title{Return counterfactual rewards estimated by a categorical reward estimator for
each observation in the supplied data and predictions}
\usage{
\method{predict_reward}{categorical_reward_estimator}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to use for estimation}

\item{X}{The features of the data.}

\item{...}{Additional arguments depending on the treatment and outcome types.
Refer to the Julia documentation for more information.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/RewardEstimation/reference/#IAI.predict_reward-Tuple\%7BCategoricalRewardEstimator\%7D}{\code{IAI.predict_reward}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{iai::predict_reward(lnr, X, treatments, outcomes, predictions)}

}
