% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict_reward}
\alias{predict_reward}
\title{Generic function for returning the counterfactual rewards estimated by a
model under each treatment}
\usage{
predict_reward(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the counterfactual rewards estimated by a
model under each treatment
}
