% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{predict_shap}
\alias{predict_shap}
\title{Calculate SHAP values for all points in the features using the learner}
\usage{
predict_shap(lnr, X)
}
\arguments{
\item{lnr}{The XGBoost learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/Heuristics/reference/#IAI.predict_shap}{\code{IAI.predict_shap}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::predict_shap(lnr, X)}

}
