% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_treatment_outcome_standard_error}
\alias{predict_treatment_outcome_standard_error}
\title{Return the standard error for the estimated quality of each treatment in the
trained model of the learner for each point in the features}
\usage{
predict_treatment_outcome_standard_error(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAIBase/reference/#IAI.predict_treatment_outcome_standard_error}{\code{IAI.predict_treatment_outcome_standard_error}}
}
\section{IAI Compatibility}{

Requires IAI version 3.2 or higher.
}

\examples{
\dontrun{iai::predict_treatment_outcome_standard_error(lnr, X)}

}
