% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_treatment_rank}
\alias{predict_treatment_rank}
\title{Return the treatments in ranked order of effectiveness for each point in the
features}
\usage{
predict_treatment_rank(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAIBase/reference/#IAI.predict_treatment_rank}{\code{IAI.predict_treatment_rank}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::predict_treatment_rank(lnr, X)}

}
