% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{print_path}
\alias{print_path}
\title{print_path}
\usage{
print_path(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{X}{The features of the data.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Print the decision path through the learner for each sample in the features
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.print_path}{\code{IAI.print_path}}
}
\examples{
\dontrun{
iai::print_path(lnr, X)
iai::print_path(lnr, X, 1)
}

}
