% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{questionnaire.optimal_feature_selection_learner}
\alias{questionnaire.optimal_feature_selection_learner}
\title{Specify an interactive questionnaire of an Optimal Feature Selection learner}
\usage{
\method{questionnaire}{optimal_feature_selection_learner}(obj, ...)
}
\arguments{
\item{obj}{The learner to visualize.}

\item{...}{Refer to the \href{https://docs.interpretable.ai/v3.2.1/IAITrees/advanced/#Advanced-Visualization-1}{Julia documentation} for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/OptimalFeatureSelection/reference/#IAI.Questionnaire-Tuple\%7BOptimalFeatureSelectionLearner\%7D}{\code{IAI.Questionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::questionnaire(lnr)}

}
