% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{questionnaire}
\alias{questionnaire}
\title{Specify an interactive questionnaire of a tree learner}
\usage{
questionnaire(lnr, ...)
}
\arguments{
\item{lnr}{The learner to visualize.}

\item{...}{Refer to the \href{https://docs.interpretable.ai/v2.1.0/IAITrees/advanced/#Advanced-Visualization-1}{Julia documentation} for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAITrees/reference/#IAI.Questionnaire}{\code{IAI.Questionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::questionnaire(lnr)}

}
