% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{questionnaire.tree_learner}
\alias{questionnaire.tree_learner}
\title{Specify an interactive questionnaire of a tree learner}
\usage{
\method{questionnaire}{tree_learner}(obj, ...)
}
\arguments{
\item{obj}{The learner to visualize.}

\item{...}{Refer to the \href{https://docs.interpretable.ai/v3.2.1/IAITrees/advanced/#Advanced-Visualization-1}{Julia documentation} for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.Questionnaire-Tuple\%7BTreeLearner\%7D}{\code{IAI.Questionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::questionnaire(lnr)}

}
