% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{rand_imputation_learner}
\alias{rand_imputation_learner}
\title{Learner for conducting random imputation}
\usage{
rand_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.RandImputationLearner}{\code{IAI.RandImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::rand_imputation_learner()}

}
