% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{random_forest_classifier}
\alias{random_forest_classifier}
\title{Learner for training random forests for classification problems}
\usage{
random_forest_classifier(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/Heuristics/reference/#IAI.RandomForestClassifier}{\code{IAI.RandomForestClassifier}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::random_forest_classifier()}

}
