% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{random_forest_regressor}
\alias{random_forest_regressor}
\title{Learner for training random forests for regression problems}
\usage{
random_forest_regressor(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/Heuristics/reference/#IAI.RandomForestRegressor}{\code{IAI.RandomForestRegressor}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::random_forest_regressor()}

}
