% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{read_json}
\alias{read_json}
\title{read_json}
\usage{
read_json(filename)
}
\arguments{
\item{filename}{The location of the JSON file.}
}
\description{
Read in a learner saved in JSON format
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.read_json}{\code{IAI.read_json}}
}
\examples{
\dontrun{lnr <- iai::read_json("out.json")}

}
