% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{refit_leaves}
\alias{refit_leaves}
\title{Refit the models in the leaves of a trained learner using the supplied data}
\usage{
refit_leaves(lnr, ...)
}
\arguments{
\item{lnr}{The learner to refit}

\item{...}{Refer to the Julia documentation for available parameters}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/OptimalTrees/reference/#IAI.refit_leaves!}{\code{IAI.refit_leaves!}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{iai::refit_leaves(lnr, ...)}

}
