% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{release_license}
\alias{release_license}
\title{Release any IAI license held by the current session.}
\usage{
release_license()
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/installation/#IAI.release_license}{\code{IAI.release_license}}
}
\section{IAI Compatibility}{

Requires IAI version 3.1 or higher.
}

\examples{
\dontrun{iai::release_license()}

}
