% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{reset_display_label}
\alias{reset_display_label}
\title{reset_display_label}
\usage{
reset_display_label(lnr)
}
\arguments{
\item{lnr}{The learner to modify.}
}
\description{
Reset the predicted probability displayed to be that of the predicted label
when visualizing a learner
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.reset_display_label!}{\code{IAI.reset_display_label!}}
}
\examples{
\dontrun{iai::reset_display_label(lnr)}

}
