% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{resume_from_checkpoint}
\alias{resume_from_checkpoint}
\title{Resume training from a checkpoint file}
\usage{
resume_from_checkpoint(checkpoint_file)
}
\arguments{
\item{checkpoint_file}{The location of the checkpoint file.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAIBase/reference/#IAI.resume_from_checkpoint}{\code{IAI.resume_from_checkpoint}}
}
\section{IAI Compatibility}{

Requires IAI version 3.1 or higher.
}

\examples{
\dontrun{obj <- iai::resume_from_checkpoint("checkpoint.json")}

}
