% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{reward_estimator}
\alias{reward_estimator}
\title{Learner for conducting Reward Estimation}
\usage{
reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/RewardEstimation/reference/#IAI.RewardEstimator}{\code{IAI.RewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{lnr <- iai::reward_estimator()}

}
