% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{roc_curve.default}
\alias{roc_curve.default}
\title{Construct an ROC curve from predicted probabilities and true labels}
\usage{
\method{roc_curve}{default}(obj, y, positive_label = stop("`positive_label` is required"), ...)
}
\arguments{
\item{obj}{The predicted probabilities for each point in the data.}

\item{y}{The true labels of the data.}

\item{positive_label}{The label for which probability is being predicted.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAIBase/reference/#IAI.ROCCurve}{\code{IAI.ROCCurve}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::roc_curve(probs, y, positive_label=positive_label)}

}
