% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{roc_curve.learner}
\alias{roc_curve.learner}
\title{Construct an ROC curve using a trained model on the given data}
\usage{
\method{roc_curve}{learner}(obj, X, y, ...)
}
\arguments{
\item{obj}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{y}{The labels of the data.}

\item{...}{Additional arguments (unused)}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAIBase/reference/#IAI.ROCCurve}{\code{IAI.ROCCurve}}
}
\examples{
\dontrun{iai::roc_curve(lnr, X, y)}

}
