% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{roc_curve}
\alias{roc_curve}
\title{Construct an ROC curve using a trained model on the given data}
\usage{
roc_curve(lnr, X, y)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{y}{The labels of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAIBase/reference/#IAI.ROCCurve}{\code{IAI.ROCCurve}}
}
\examples{
\dontrun{iai::roc_curve(lnr, X, y)}

}
