% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{score.learner}
\alias{score.learner}
\title{Calculate the score for a model on the given data}
\usage{
\method{score}{learner}(obj, X, ...)
}
\arguments{
\item{obj}{The learner or grid to evaluate.}

\item{X}{The features of the data.}

\item{...}{Other parameters, including zero or more target vectors as
required by the problem type. Refer to the Julia documentation for
available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAIBase/reference/#IAI.score}{\code{IAI.score}}
}
\examples{
\dontrun{iai::score(obj, X, y)}

}
