% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{score}
\alias{score}
\title{Generic function for calculating scores}
\usage{
score(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for calculating scores
}
