% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_julia_seed}
\alias{set_julia_seed}
\title{set_julia_seed}
\usage{
set_julia_seed(seed)
}
\arguments{
\item{seed}{The seed to set}
}
\description{
Set the random seed in Julia
}
\details{
Julia Equivalent:
\href{https://docs.julialang.org/en/v1/stdlib/Random/index.html#Random.seed!}{\code{Random.seed!}}
}
\examples{
\dontrun{iai::set_julia_seed(1)}

}
