% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{set_params}
\alias{set_params}
\title{Set all supplied parameters on a learner}
\usage{
set_params(lnr, ...)
}
\arguments{
\item{lnr}{The learner to modify.}

\item{...}{The parameters to set on the learner.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAIBase/reference/#IAI.set_params!}{\code{IAI.set_params!}}
}
\examples{
\dontrun{iai::set_params(lnr, random_seed = 1)}

}
