% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{set_reward_kernel_bandwidth}
\alias{set_reward_kernel_bandwidth}
\title{Save a new reward kernel bandwidth inside a learner, and return new reward
predictions generated using this bandwidth for the original data used to
train the learner.}
\usage{
set_reward_kernel_bandwidth(lnr, ...)
}
\arguments{
\item{lnr}{The learner to modify}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/RewardEstimation/reference/#IAI.set_reward_kernel_bandwidth!}{\code{IAI.set_reward_kernel_bandwidth!}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::set_reward_kernel_bandwidth(lnr, ...)}

}
