% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{set_rich_output_param}
\alias{set_rich_output_param}
\title{set_rich_output_param}
\usage{
set_rich_output_param(key, value)
}
\arguments{
\item{key}{The parameter to set.}

\item{value}{The value to set}
}
\description{
Sets a global rich output parameter
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.set_rich_output_param!}{\code{IAI.set_rich_output_param!}}
}
\examples{
\dontrun{iai::set_rich_output_param("simple_layout", TRUE)}

}
