% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{show_in_browser.abstract_visualization}
\alias{show_in_browser.abstract_visualization}
\title{Show interactive visualization of an object in the default browser}
\usage{
\method{show_in_browser}{abstract_visualization}(obj, ...)
}
\arguments{
\item{obj}{The object to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAIBase/reference/#IAI.show_in_browser-Tuple\%7BAbstractVisualization\%7D}{\code{IAI.show_in_browser}}
}
\examples{
\dontrun{iai::show_in_browser(lnr)}

}
