% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{show_in_browser.roc_curve}
\alias{show_in_browser.roc_curve}
\title{Show interactive visualization of a \code{\link{roc_curve}} in the default
browser}
\usage{
\method{show_in_browser}{roc_curve}(obj, ...)
}
\arguments{
\item{obj}{The curve to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.show_in_browser-Tuple\%7BROCCurve\%7D}{\code{IAI.show_in_browser}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::show_in_browser(curve)}

}
