% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{show_in_browser.tree_learner}
\alias{show_in_browser.tree_learner}
\title{Show interactive tree visualization of a tree learner in the default browser}
\usage{
\method{show_in_browser}{tree_learner}(obj, ...)
}
\arguments{
\item{obj}{The learner or grid to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAITrees/reference/#IAI.show_in_browser-Tuple\%7BTreeLearner\%7D}{\code{IAI.show_in_browser}}
}
\section{IAI Compatibility}{

Showing a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::show_in_browser(lnr)}

}
