% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{show_questionnaire.optimal_feature_selection_learner}
\alias{show_questionnaire.optimal_feature_selection_learner}
\title{Show an interactive questionnaire based on an Optimal Feature Selection
learner in default browser}
\usage{
\method{show_questionnaire}{optimal_feature_selection_learner}(obj, ...)
}
\arguments{
\item{obj}{The learner or grid to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/OptimalFeatureSelection/reference/#IAI.show_questionnaire-Tuple\%7BOptimalFeatureSelectionLearner\%7D}{\code{IAI.show_questionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::show_questionnaire(lnr)}

}
