% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{show_questionnaire.tree_learner}
\alias{show_questionnaire.tree_learner}
\title{Show an interactive questionnaire based on a tree learner in default browser}
\usage{
\method{show_questionnaire}{tree_learner}(obj, ...)
}
\arguments{
\item{obj}{The learner or grid to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/IAITrees/reference/#IAI.show_questionnaire-Tuple\%7BTreeLearner\%7D}{\code{IAI.show_questionnaire}}
}
\section{IAI Compatibility}{

Showing a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::show_questionnaire(lnr)}

}
