% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{single_knn_imputation_learner}
\alias{single_knn_imputation_learner}
\title{Learner for conducting heuristic k-NN imputation}
\usage{
single_knn_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.SingleKNNImputationLearner}{\code{IAI.SingleKNNImputationLearner}}
}
\examples{
\dontrun{lnr <- iai::single_knn_imputation_learner()}

}
