% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{stability_analysis}
\alias{stability_analysis}
\title{Conduct a stability analysis of the trees in a tree learner}
\usage{
stability_analysis(lnr, ...)
}
\arguments{
\item{lnr}{The original learner}

\item{...}{Additional arguments (refer to Julia documentation)}
}
\description{
Refer to the
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/stability/#Tree-Stability-1}{documentation on tree stability}
for more information.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.StabilityAnalysis}{\code{StabilityAnalysis}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::stability_analysis(lnr, ...)}

}
