% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{transform}
\alias{transform}
\title{transform}
\usage{
transform(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for imputation}

\item{X}{The features of the data.}
}
\description{
Impute missing values in a dataframe using a fitted imputation model
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.transform}{\code{IAI.transform}}
}
\examples{
\dontrun{iai::transform(lnr, X)}

}
