% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{transform_and_expand}
\alias{transform_and_expand}
\title{Transform features with a trained imputation learner and create adaptive
indicator features to encode the missing pattern}
\usage{
transform_and_expand(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner to use for imputation.}

\item{X}{The dataframe in which to impute missing values.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/OptImpute/reference/#IAI.transform_and_expand}{\code{IAI.transform_and_expand}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{lnr <- iai::transform_and_expand(lnr, X, type = "finite")}

}
