% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{tree_plot}
\alias{tree_plot}
\title{Specify an interactive tree visualization of a tree learner}
\usage{
tree_plot(lnr, ...)
}
\arguments{
\item{lnr}{The learner to visualize.}

\item{...}{Refer to the \href{https://docs.interpretable.ai/v2.2.0/IAITrees/advanced/#Advanced-Visualization-1}{Julia documentation on advanced tree visualization} for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.TreePlot}{\code{IAI.TreePlot}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::tree_plot(lnr)}

}
