% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{variable_importance.learner}
\alias{variable_importance.learner}
\title{Generate a ranking of the variables in a learner according to their
importance during training. The results are normalized so that they sum to
one.}
\usage{
\method{variable_importance}{learner}(obj, ...)
}
\arguments{
\item{obj}{The learner to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.variable_importance-Tuple\%7BLearner\%7D}{\code{IAI.variable_importance}}
}
\examples{
\dontrun{iai::variable_importance(lnr, ...)}

}
