% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{variable_importance}
\alias{variable_importance}
\title{Generic function for calculating variable importance}
\usage{
variable_importance(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for calculating variable importance
}
