% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{write_booster}
\alias{write_booster}
\title{Write the internal booster saved in the learner to file}
\usage{
write_booster(filename, lnr)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The XGBoost learner with the booster to output.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/Heuristics/reference/#IAI.write_booster}{\code{IAI.write_booster}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::write_booster(file.path(tempdir(), "out.json"), lnr)}

}
