% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_dot}
\alias{write_dot}
\title{Output a learner in
\href{http://www.graphviz.org/content/dot-language/}{.dot format}}
\usage{
write_dot(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.write_dot}{\code{IAI.write_dot}}
}
\examples{
\dontrun{iai::write_dot(file.path(tempdir(), "tree.dot"), lnr)}

}
