% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{write_html.abstract_visualization}
\alias{write_html.abstract_visualization}
\title{Output an object as an interactive browser visualization in HTML format}
\usage{
\method{write_html}{abstract_visualization}(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The object to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAIBase/reference/#IAI.write_html-Tuple\%7BAny\%2C\%20AbstractVisualization\%7D}{\code{IAI.write_html}}
}
\examples{
\dontrun{iai::write_html(file.path(tempdir(), "out.html"), lnr)}

}
