% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{write_html.roc_curve}
\alias{write_html.roc_curve}
\title{Output an ROC curve as an interactive browser visualization in HTML format}
\usage{
\method{write_html}{roc_curve}(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The curve to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.write_html-Tuple\%7BAny\%2C\%20ROCCurve\%7D}{\code{IAI.write_html}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{iai::write_html(file.path(tempdir(), "roc.html"), lnr)}

}
