% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_html.tree_learner}
\alias{write_html.tree_learner}
\title{Output a tree learner as an interactive browser visualization in HTML format}
\usage{
\method{write_html}{tree_learner}(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The learner or grid to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.write_html-Tuple\%7BAny\%2C\%20TreeLearner\%7D}{\code{IAI.write_html}}
}
\section{IAI Compatibility}{

Outputting a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::write_html(file.path(tempdir(), "tree.html"), lnr)}

}
