% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{write_json}
\alias{write_json}
\title{write_json}
\usage{
write_json(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Output a learner in JSON format
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.write_json}{\code{IAI.write_json}}
}
\examples{
\dontrun{iai::write_json(file.path(tempdir(), "out.json"), lnr)}

}
