% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_pdf}
\alias{write_pdf}
\title{Output a learner as a PDF image}
\usage{
write_pdf(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.write_pdf}{\code{IAI.write_pdf}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::write_pdf(file.path(tempdir(), "tree.pdf"), lnr)}

}
