% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalfeatureselection.R
\name{write_questionnaire.optimal_feature_selection_learner}
\alias{write_questionnaire.optimal_feature_selection_learner}
\title{Output an Optimal Feature Selection learner as an interactive questionnaire
in HTML format}
\usage{
\method{write_questionnaire}{optimal_feature_selection_learner}(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The learner or grid to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/OptimalFeatureSelection/reference/#IAI.write_questionnaire-Tuple\%7BAny\%2C\%20OptimalFeatureSelectionLearner\%7D}{\code{IAI.write_questionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::write_questionnaire(file.path(tempdir(), "questionnaire.html"), lnr)}

}
