% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_questionnaire.tree_learner}
\alias{write_questionnaire.tree_learner}
\title{Output a tree learner as an interactive questionnaire in HTML format}
\usage{
\method{write_questionnaire}{tree_learner}(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The learner or grid to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/IAITrees/reference/#IAI.write_questionnaire-Tuple\%7BAny\%2C\%20TreeLearner\%7D}{\code{IAI.write_questionnaire}}
}
\section{IAI Compatibility}{

Outputting a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::write_questionnaire(file.path(tempdir(), "questionnaire.html"), lnr)}

}
