% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_svg}
\alias{write_svg}
\title{Output a learner as a SVG image}
\usage{
write_svg(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Before using this function, either run \code{\link{load_graphviz}} or ensure
that \href{https://graphviz.org/}{Graphviz} is installed and on the system
\code{PATH}
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.write_svg}{\code{IAI.write_svg}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{iai::write_svg(file.path(tempdir(), "tree.svg"), lnr)}

}
