% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{xgboost_classifier}
\alias{xgboost_classifier}
\title{Learner for training XGBoost models for classification problems}
\usage{
xgboost_classifier(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/Heuristics/reference/#IAI.XGBoostClassifier}{\code{IAI.XGBoostClassifier}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::xgboost_classifier()}

}
