% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{xgboost_survival_learner}
\alias{xgboost_survival_learner}
\title{Learner for training XGBoost models for survival problems}
\usage{
xgboost_survival_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/Heuristics/reference/#IAI.XGBoostSurvivalLearner}{\code{IAI.XGBoostSurvivalLearner}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{lnr <- iai::xgboost_survival_learner()}

}
