% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{zero_imputation_learner}
\alias{zero_imputation_learner}
\title{Learner for conducting zero-imputation}
\usage{
zero_imputation_learner(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/OptImpute/reference/#IAI.ZeroImputationLearner}{\code{IAI.ZeroImputationLearner}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{lnr <- iai::zero_imputation_learner()}

}
