% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{categorical_reward_estimator}
\alias{categorical_reward_estimator}
\title{Learner for conducting reward estimation with categorical treatments}
\usage{
categorical_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/RewardEstimation/reference/#IAI.CategoricalRewardEstimator}{\code{IAI.CategoricalRewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{lnr <- iai::categorical_reward_estimator()}

}
