% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{clone}
\alias{clone}
\title{Return an unfitted copy of a learner with the same parameters}
\usage{
clone(lnr)
}
\arguments{
\item{lnr}{The learner to copy.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAIBase/reference/#IAI.clone}{\code{IAI.clone}}
}
\examples{
\dontrun{new_lnr <- iai::clone(lnr)}

}
