% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{fit}
\alias{fit}
\title{Generic function for fitting a learner.}
\usage{
fit(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for fitting a learner.
}
