% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{fit_predict}
\alias{fit_predict}
\title{Generic function for fitting a reward estimator on features,
treatments and returning predicted counterfactual rewards and scores of the
internal estimators.}
\usage{
fit_predict(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/RewardEstimation/reference/#IAI.fit_predict\%21-Tuple\%7BRewardEstimator\%7D}{\code{IAI.fit_predict!}}
}
