% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_classification_label}
\alias{get_classification_label}
\title{Return the predicted label at a node of a tree}
\usage{
get_classification_label(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_classification_label}{\code{IAI.get_classification_label}}
}
\examples{
\dontrun{iai::get_classification_label(lnr, 1)}

}
