% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_classification_proba}
\alias{get_classification_proba}
\title{get_classification_proba}
\usage{
get_classification_proba(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Return the predicted probabilities of class membership at a node of a tree
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_classification_proba}{\code{IAI.get_classification_proba}}
}
\examples{
\dontrun{iai::get_classification_proba(lnr, 1)}

}
