% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_classification_proba}
\alias{get_classification_proba}
\title{Return the predicted probabilities of class membership at a node of a tree}
\usage{
get_classification_proba(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.get_classification_proba}{\code{IAI.get_classification_proba}}
}
\examples{
\dontrun{iai::get_classification_proba(lnr, 1)}

}
