% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_grid_results}
\alias{get_grid_results}
\title{get_grid_results}
\usage{
get_grid_results(grid)
}
\arguments{
\item{grid}{The grid search to query.}
}
\description{
Return a summary of the results from the grid search
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.get_grid_results}{\code{IAI.get_grid_results}}
}
\examples{
\dontrun{iai::get_grid_results(grid)}

}
