% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_learner}
\alias{get_learner}
\title{get_learner}
\usage{
get_learner(grid)
}
\arguments{
\item{grid}{The grid to query.}
}
\description{
Return the fitted learner using the best parameter combination from a grid
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.get_learner}{\code{IAI.get_learner}}
}
\examples{
\dontrun{lnr <- iai::get_learner(grid)}

}
