% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{get_num_fits}
\alias{get_num_fits}
\title{Return the number of fits along the path in the trained learner}
\usage{
get_num_fits(lnr)
}
\arguments{
\item{lnr}{The GLMNet learner to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/Heuristics/reference/#IAI.get_num_fits}{\code{IAI.get_num_fits}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::get_num_fits(lnr)}

}
