% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_num_nodes}
\alias{get_num_nodes}
\title{get_num_nodes}
\usage{
get_num_nodes(lnr)
}
\arguments{
\item{lnr}{The learner or grid to query.}
}
\description{
Return the number of nodes in a trained learner
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_num_nodes}{\code{IAI.get_num_nodes}}
}
\examples{
\dontrun{iai::get_num_nodes(lnr)}

}
