% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_num_samples}
\alias{get_num_samples}
\title{Get the number of training points contained in a node of a tree}
\usage{
get_num_samples(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.get_num_samples}{\code{IAI.get_num_samples}}
}
\examples{
\dontrun{iai::get_num_samples(lnr, 1)}

}
