% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_prescription_treatment_rank}
\alias{get_prescription_treatment_rank}
\title{Return the treatments ordered from most effective to least effective at a
node of a tree}
\usage{
get_prescription_treatment_rank(lnr, node_index, ...)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.get_prescription_treatment_rank}{\code{IAI.get_prescription_treatment_rank}}
}
\examples{
\dontrun{iai::get_prescription_treatment_rank(lnr, 1)}

}
