% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_regression_weights}
\alias{get_regression_weights}
\title{Generic function for returning the weights for each feature in the
regression prediction at a node of a tree}
\usage{
get_regression_weights(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the weights for each feature in the
regression prediction at a node of a tree
}
