% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{get_rich_output_params}
\alias{get_rich_output_params}
\title{get_rich_output_params}
\usage{
get_rich_output_params()
}
\description{
Return the current global rich output parameter settings
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.get_rich_output_params}{\code{IAI.get_rich_output_params}}
}
\examples{
\dontrun{iai::get_rich_output_params()}

}
