% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_split_threshold}
\alias{get_split_threshold}
\title{Return the threshold used in the split at a node of a tree}
\usage{
get_split_threshold(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.0/IAITrees/reference/#IAI.get_split_threshold}{\code{IAI.get_split_threshold}}
}
\examples{
\dontrun{iai::get_split_threshold(lnr, 1)}

}
