% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_split_weights}
\alias{get_split_weights}
\title{Return the weights for numeric and categoric features used in the hyperplane
split at a node of a tree}
\usage{
get_split_weights(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAITrees/reference/#IAI.get_split_weights}{\code{IAI.get_split_weights}}
}
\examples{
\dontrun{iai::get_split_weights(lnr, 1)}

}
