% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_survival_curve}
\alias{get_survival_curve}
\title{get_survival_curve}
\usage{
get_survival_curve(lnr, node_index)
}
\arguments{
\item{lnr}{The learner or grid to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Return the survival curve at a node of a tree
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.get_survival_curve}{\code{IAI.get_survival_curve}}
}
\examples{
\dontrun{iai::get_survival_curve(lnr, 1)}

}
