% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{get_upper_child}
\alias{get_upper_child}
\title{Get the index of the upper child at a split node of a tree}
\usage{
get_upper_child(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAITrees/reference/#IAI.get_upper_child}{\code{IAI.get_upper_child}}
}
\examples{
\dontrun{iai::get_upper_child(lnr, 1)}

}
