% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{iai_setup}
\alias{iai_setup}
\title{Initialize Julia and the IAI package.}
\usage{
iai_setup(...)
}
\arguments{
\item{...}{All parameters are passed through to
\href{https://www.rdocumentation.org/packages/JuliaCall/topics/julia_setup}{\code{JuliaCall::julia_setup}}}
}
\description{
This function is called automatically with default parameters the first time
any `iai` function is used in an R session. If custom parameters for Julia
setup are required, this function must be called in every R session before
calling other `iai` functions.
}
\examples{
\dontrun{iai::iai_setup()}

}
