% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{iai_setup}
\alias{iai_setup}
\title{Initialize Julia and the IAI package.}
\usage{
iai_setup(...)
}
\arguments{
\item{...}{All parameters are passed through to
\href{https://www.rdocumentation.org/packages/JuliaCall/topics/julia_setup}{\code{JuliaCall::julia_setup}}}
}
\description{
This needs to be done in every R session before calling `iai` functions
}
\examples{
\dontrun{iai::iai_setup()}

}
