% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{impute_cv}
\alias{impute_cv}
\title{impute_cv}
\usage{
impute_cv(X, ...)
}
\arguments{
\item{X}{The dataframe in which to impute missing values.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Impute missing values using cross validation
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/OptImpute/stable/reference/#IAI.impute_cv}{\code{IAI.impute_cv}}
}
\examples{
\donttest{
X <- iris
X[1, 1] <- NA
iai::iai_setup()
iai::impute_cv(X, list(method = c("opt_knn", "opt_tree")))
}
}
