% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimpute.R
\name{impute_cv}
\alias{impute_cv}
\title{Impute missing values using cross validation}
\usage{
impute_cv(X, ...)
}
\arguments{
\item{X}{The dataframe in which to impute missing values.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.1.1/OptImpute/reference/#IAI.impute_cv}{\code{IAI.impute_cv}}
}
\details{
This function was deprecated in iai 1.7.0. This is for consistency with the
IAI v3.0.0 Julia release.
}
\examples{
\dontrun{
X <- iris
X[1, 1] <- NA
iai::impute_cv(X, list(method = c("opt_knn", "opt_tree")))
}
}
