% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{is_leaf}
\alias{is_leaf}
\title{Check if a node of a tree is a leaf}
\usage{
is_leaf(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAITrees/reference/#IAI.is_leaf}{\code{IAI.is_leaf}}
}
\examples{
\dontrun{iai::is_leaf(lnr, 1)}

}
