% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{is_parallel_split}
\alias{is_parallel_split}
\title{Check if a node of a tree applies a parallel split}
\usage{
is_parallel_split(lnr, node_index)
}
\arguments{
\item{lnr}{The learner to query.}

\item{node_index}{The node in the tree to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAITrees/reference/#IAI.is_parallel_split}{\code{IAI.is_parallel_split}}
}
\examples{
\dontrun{iai::is_parallel_split(lnr, 1)}

}
