% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{multi_questionnaire.default}
\alias{multi_questionnaire.default}
\title{Construct an interactive questionnaire from multiple specified learners}
\usage{
\method{multi_questionnaire}{default}(obj, ...)
}
\arguments{
\item{obj}{The questions to visualize. Refer to the \href{https://docs.interpretable.ai/v3.2.2/IAITrees/visualization/#multivis-1}{Julia documentation on multi-learner visualizations} for more information.}

\item{...}{Additional arguments (unused)}
}
\description{
Refer to the
\href{https://docs.interpretable.ai/v3.2.2/IAI-R/julia/#R-Interactive-Visualizations-1}{documentation on advanced tree visualization}
for more information.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAIBase/reference/#IAI.MultiQuestionnaire-Tuple\%7BPair\%7D}{\code{IAI.MultiQuestionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{
iai::multi_questionnaire(list("Questionnaire for" = list(
   "first learner" = lnr1,
   "second learner" = lnr2
)))
}

}
