% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{multi_questionnaire}
\alias{multi_questionnaire}
\title{Construct an interactive questionnaire using multiple tree learners as
specified by questions}
\usage{
multi_questionnaire(questions)
}
\arguments{
\item{questions}{The questions to visualize. Refer to the \href{https://docs.interpretable.ai/v1.2.0/IAITrees/visualization/#multivis-1}{Julia documentation on multi-learner visualizations} for more information.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAITrees/reference/#IAI.MultiQuestionnaire}{\code{IAI.MultiQuestionnaire}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{
iai::multi_questionnaire(list("Questionnaire for" = list(
   "first learner" = lnr1,
   "second learner" = lnr2
)))
}

}
