% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{multi_tree_plot.default}
\alias{multi_tree_plot.default}
\title{Construct an interactive tree visualization of multiple tree learners as
specified by questions}
\usage{
\method{multi_tree_plot}{default}(obj, ...)
}
\arguments{
\item{obj}{The questions to visualize. Refer to the \href{https://docs.interpretable.ai/v3.2.1/IAITrees/visualization/#multivis-1}{Julia documentation on multi-learner visualizations} for more information.}

\item{...}{Additional arguments (unused)}
}
\description{
Refer to the
\href{https://docs.interpretable.ai/v3.2.1/IAI-R/julia/#R-Interactive-Visualizations-1}{documentation on advanced tree visualization}
for more information.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.MultiTreePlot-Tuple\%7BPair\%7D}{\code{IAI.MultiTreePlot}}
}
\section{IAI Compatibility}{

Requires IAI version 1.1 or higher.
}

\examples{
\dontrun{
iai::multi_tree_plot(list("Visualizing" = list(
   "first learner" = lnr1,
   "second learner" = lnr2
)))
}

}
