% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{multi_tree_plot}
\alias{multi_tree_plot}
\title{Generic function for constructing an interactive tree visualization of
multiple tree learners}
\usage{
multi_tree_plot(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAITrees/reference/#IAI.MultiTreePlot}{\code{IAI.MultiTreePlot}}
}
