% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{numeric_reward_estimator}
\alias{numeric_reward_estimator}
\title{Learner for conducting reward estimation with numeric treatments}
\usage{
numeric_reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/RewardEstimation/reference/#IAI.NumericRewardEstimator}{\code{IAI.NumericRewardEstimator}}
}
\section{IAI Compatibility}{

Requires IAI version 2.1 or higher.
}

\examples{
\dontrun{lnr <- iai::numeric_reward_estimator()}

}
