% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_regressor}
\alias{optimal_tree_regressor}
\title{Learner for training Optimal Regression Trees}
\usage{
optimal_tree_regressor(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/OptimalTrees/reference/#IAI.OptimalTreeRegressor}{\code{IAI.OptimalTreeRegressor}}
}
\examples{
\dontrun{lnr <- iai::optimal_tree_regressor()}

}
