% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaltrees.R
\name{optimal_tree_survivor}
\alias{optimal_tree_survivor}
\title{Learner for training Optimal Survival Trees}
\usage{
optimal_tree_survivor(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
This function was deprecated and renamed to \code{\link[=optimal_tree_survival_learner]{optimal_tree_survival_learner()}}
in iai 1.3.0. This is for consistency with the IAI v2.0.0 Julia release.
}
\examples{
\dontrun{lnr <- iai::optimal_tree_survivor()}

}
