% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict}
\alias{predict}
\title{Return the predictions made by the model for each point in the features}
\usage{
predict(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAIBase/reference/#IAI.predict}{\code{IAI.predict}}
}
\examples{
\dontrun{iai::predict(lnr, X)}

}
