% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_expected_survival_time}
\alias{predict_expected_survival_time}
\title{Return the expected survival time estimate made by a model for each
point in the features.}
\usage{
predict_expected_survival_time(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAIBase/reference/#IAI.predict_expected_survival_time}{\code{IAI.predict_expected_survival_time}}
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::predict_expected_survival_time(lnr, X)}

}
