% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_hazard}
\alias{predict_hazard}
\title{Return the fitted hazard coefficient estimate made by a model for each
point in the features.}
\usage{
predict_hazard(lnr, X)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}
}
\description{
A higher hazard coefficient estimate corresponds to a smaller predicted
survival time.
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAIBase/reference/#IAI.predict_hazard}{\code{IAI.predict_hazard}}
}
\section{IAI Compatibility}{

Requires IAI version 1.2 or higher.
}

\examples{
\dontrun{iai::predict_hazard(lnr, X)}

}
