% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict_outcomes}
\alias{predict_outcomes}
\title{Generic function for returning the outcomes predicted by a model under each
treatment}
\usage{
predict_outcomes(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the outcomes predicted by a model under each
treatment
}
