% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{predict_outcomes}
\alias{predict_outcomes}
\title{Return the predicted outcome for each treatment made by a model for each
point in the features}
\usage{
predict_outcomes(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{...}{For policy learners only, the reward matrix.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAIBase/reference/#IAIBase-API-Reference-1}{\code{IAI.predict_outcomes}}
(for prescription or policy learners as appropriate)
}
\section{IAI Compatibility}{

Requires IAI version 2.0 or higher for policy learners.
}

\examples{
\dontrun{iai::predict_outcomes(lnr, X, ...)}

}
