% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict_proba}
\alias{predict_proba}
\title{Generic function for returning the probabilities of class membership
predicted by a model}
\usage{
predict_proba(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for returning the probabilities of class membership
predicted by a model
}
