% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{predict_reward}
\alias{predict_reward}
\title{Return counterfactual rewards estimated using learner parameters for each
observation in the supplied data and predictions}
\usage{
predict_reward(lnr, ...)
}
\arguments{
\item{lnr}{The learner or grid to use for estimation}

\item{...}{Additional arguments depending on the treatment and outcome types.
Refer to the Julia documentation for more information.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/RewardEstimation/reference/#IAI.predict_reward}{\code{IAI.predict_reward}}
}
\section{IAI Compatibility}{

Requires IAI version 3.0 or higher.
}

\examples{
\dontrun{iai::predict_reward(lnr, treatments, outcomes, predictions)}

}
