% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{read_json}
\alias{read_json}
\title{Read in a learner or grid saved in JSON format}
\usage{
read_json(filename)
}
\arguments{
\item{filename}{The location of the JSON file.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAIBase/reference/#IAI.read_json}{\code{IAI.read_json}}
}
\examples{
\dontrun{obj <- iai::read_json("out.json")}

}
