% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{reward_estimator}
\alias{reward_estimator}
\title{Learner for conducting reward estimation with categorical treatments}
\usage{
reward_estimator(...)
}
\arguments{
\item{...}{Use keyword arguments to set parameters on the resulting learner.
Refer to the Julia documentation for available parameters.}
}
\description{
This function was deprecated and renamed to \code{\link[=categorical_reward_estimator]{categorical_reward_estimator()}}
in iai 1.4.0. This is for consistency with the IAI v2.1.0 Julia release.
}
\details{
This deprecation is no longer supported as of the IAI v3 release.
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or lower.
}

\examples{
\dontrun{lnr <- iai::reward_estimator()}

}
