% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{roc_curve}
\alias{roc_curve}
\title{Generic function for constructing an ROC curve}
\usage{
roc_curve(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAIBase/reference/#IAI.ROCCurve}{\code{IAI.ROCCurve}}
}
