% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{roc_curve}
\alias{roc_curve}
\title{roc_curve}
\usage{
roc_curve(lnr, X, y)
}
\arguments{
\item{lnr}{The learner or grid to use for prediction.}

\item{X}{The features of the data.}

\item{y}{The labels of the data.}
}
\description{
Construct an ROC curve using a trained model on the given data
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.ROCCurve}{\code{IAI.ROCCurve}}
}
\examples{
\dontrun{iai::roc_curve(lnr, X, y)}

}
