% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{score}
\alias{score}
\title{Generic function for calculating scores}
\usage{
score(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.0.0/IAIBase/reference/#IAI.score}{\code{IAI.score}}
}
