% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{score}
\alias{score}
\title{score}
\usage{
score(lnr, X, ...)
}
\arguments{
\item{lnr}{The learner or grid to evaluate.}

\item{X}{The features of the data.}

\item{...}{Other parameters, including zero or more target vectors as
required by the problem type. Refer to the Julia documentation for
available parameters.}
}
\description{
Calculate the score for a model on the given data
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAIBase/stable/reference/#IAI.score}{\code{IAI.score}}
}
\examples{
\dontrun{iai::score(lnr, X, y)}

}
