% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{set_display_label}
\alias{set_display_label}
\title{Show the probability of a specified label when visualizing a learner}
\usage{
set_display_label(lnr, display_label)
}
\arguments{
\item{lnr}{The learner to modify.}

\item{display_label}{The label for which to show probabilities.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/IAITrees/reference/#IAI.set_display_label!}{\code{IAI.set_display_label!}}
}
\examples{
\dontrun{iai::set_display_label(lnr, "A")}

}
