% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{show_in_browser}
\alias{show_in_browser}
\title{Generic function for showing interactive visualization in browser}
\usage{
show_in_browser(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for showing interactive visualization in browser
}
