% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{show_in_browser}
\alias{show_in_browser}
\title{Show interactive visualization of an object (such as a learner or curve) in
the default browser}
\usage{
show_in_browser(obj, ...)
}
\arguments{
\item{obj}{The object to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.0.0/IAIBase/reference/#IAI.show_in_browser}{\code{IAI.show_in_browser}}
}
\section{IAI Compatibility}{

Showing a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::show_in_browser(lnr)}

}
