% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{show_questionnaire}
\alias{show_questionnaire}
\title{Generic function for showing interactive questionnaire in browser}
\usage{
show_questionnaire(obj, ...)
}
\arguments{
\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for showing interactive questionnaire in browser
}
