% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{show_questionnaire}
\alias{show_questionnaire}
\title{show_questionnaire}
\usage{
show_questionnaire(lnr, ...)
}
\arguments{
\item{lnr}{The learner to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Show an interactive questionnaire based on a learner in default browser
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.show_questionnaire}{\code{IAI.show_questionnaire}}
}
\examples{
\dontrun{iai::show_questionnaire(lnr)}

}
