% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{show_questionnaire}
\alias{show_questionnaire}
\title{Show an interactive questionnaire based on a learner in default browser}
\usage{
show_questionnaire(lnr, ...)
}
\arguments{
\item{lnr}{The learner or grid to visualize.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAITrees/reference/#IAI.show_questionnaire}{\code{IAI.show_questionnaire}}
}
\section{IAI Compatibility}{

Showing a grid search requires IAI version 2.0 or higher.
}

\examples{
\dontrun{iai::show_questionnaire(lnr)}

}
