% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardestimation.R
\name{tune_reward_kernel_bandwidth}
\alias{tune_reward_kernel_bandwidth}
\title{Conduct the reward kernel bandwidth tuning procedure for a range of starting
bandwidths and return the final tuned values.}
\usage{
tune_reward_kernel_bandwidth(lnr, ...)
}
\arguments{
\item{lnr}{The learner to use for tuning the bandwidth}

\item{...}{Refer to the Julia documentation for other parameters}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.1/RewardEstimation/reference/#IAI.tune_reward_kernel_bandwidth}{\code{IAI.tune_reward_kernel_bandwidth}}
}
\section{IAI Compatibility}{

Requires IAI version 2.2 or higher.
}

\examples{
\dontrun{iai::tune_reward_kernel_bandwidth(lnr, ...)}

}
