% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{variable_importance}
\alias{variable_importance}
\title{variable_importance}
\usage{
variable_importance(lnr)
}
\arguments{
\item{lnr}{The learner or grid to query.}
}
\description{
Generate a ranking of the variables in the learner according to their
importance when training the trees
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/IAITrees/stable/reference/#IAI.variable_importance}{\code{IAI.variable_importance}}
}
\examples{
\dontrun{iai::variable_importance(lnr)}

}
