% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{variable_importance}
\alias{variable_importance}
\title{Generate a ranking of the variables in the learner according to their
importance during training. The results are normalized so that they sum to
one.}
\usage{
variable_importance(lnr)
}
\arguments{
\item{lnr}{The learner to query.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.2.0/IAIBase/reference/#IAI.variable_importance}{\code{IAI.variable_importance}}
}
\examples{
\dontrun{iai::variable_importance(lnr)}

}
