% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{write_html}
\alias{write_html}
\title{Generic function for writing interactive visualization to file}
\usage{
write_html(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The object controlling which method is used}

\item{...}{Arguments depending on the specific method used}
}
\description{
Generic function for writing interactive visualization to file
}
