% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaibase.R
\name{write_json}
\alias{write_json}
\title{Output a learner or grid in JSON format}
\usage{
write_json(filename, obj, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{obj}{The learner or grid to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v2.1.0/IAIBase/reference/#IAI.write_json}{\code{IAI.write_json}}
}
\examples{
\dontrun{iai::write_json(file.path(tempdir(), "out.json"), obj)}

}
