% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_png}
\alias{write_png}
\title{Output a learner as a PNG image}
\usage{
write_png(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Before using this function, either run \code{\link{load_graphviz}} or ensure
that \href{https://graphviz.org/}{Graphviz} is installed and on the system
\code{PATH}
}
\details{
Julia Equivalent:
\href{https://docs.interpretable.ai/v3.2.2/IAITrees/reference/#IAI.write_png}{\code{IAI.write_png}}
}
\examples{
\dontrun{iai::write_png(file.path(tempdir(), "tree.png"), lnr)}

}
