% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaitrees.R
\name{write_questionnaire}
\alias{write_questionnaire}
\title{Output a learner as an interactive questionnaire in HTML format}
\usage{
write_questionnaire(filename, lnr, ...)
}
\arguments{
\item{filename}{Where to save the output.}

\item{lnr}{The learner to output.}

\item{...}{Refer to the Julia documentation for available parameters.}
}
\description{
Julia Equivalent:
\href{https://docs.interpretable.ai/v1.2.0/IAITrees/reference/#IAI.write_questionnaire}{\code{IAI.write_questionnaire}}
}
\examples{
\dontrun{iai::write_questionnaire(file.path(tempdir(), "questionnaire.html"), lnr)}

}
